adapted to smp

+++ rdmo_orcid/handlers.py
@@ -38,6 +38,17 @@ def value_handler(sender, request=None, instance=None, **kwargs):
                 data = response.json()
             except (requests.exceptions.RequestException, requests.exceptions.HTTPError):
                 return
+            
+            if 'orcid_id' in attribute_map:
+                Value.objects.update_or_create(
+                    project=instance.project,
+                    attribute=Attribute.objects.get(uri=attribute_map['orcid_id']),
+                    set_prefix=instance.set_prefix,
+                    set_index=instance.set_index,
+                    defaults={
+                        'text': dpath.get(data, '/orcid-identifier/uri')
+                    }
+                )
 
             if 'given_name' in attribute_map:
                 Value.objects.update_or_create(
@@ -61,32 +72,67 @@ def value_handler(sender, request=None, instance=None, **kwargs):
                     }
                 )
 
-            if 'affiliation' in attribute_map:
-                attribute = Attribute.objects.get(uri=attribute_map['affiliation'])
-
-                affiliations = []
+            if 'affiliation' in attribute_map or 'role' in attribute_map:
+                affiliations = {}
+                roles = []
                 for affiliation in dpath.get(data, '/activities-summary/employments/affiliation-group'):
                     for summaries in affiliation.get('summaries'):
                         if dpath.get(summaries, '/employment-summary/end-date') is None:
-                            affiliations.append(dpath.get(summaries, '/employment-summary/organization/name'))
+                            a = dpath.get(summaries, '/employment-summary/organization/name')
+                            role = dpath.get(summaries, '/employment-summary/role-title')
+                            a_roles = affiliations.get(a, [])
+                            if role != None:
+                                a_roles.append(role)
+                                roles.append(role)
 
-                for collection_index, affiliation in enumerate(affiliations):
-                    Value.objects.update_or_create(
+                            affiliations[a] = a_roles
+
+                if 'role' in attribute_map:
+                    attribute = Attribute.objects.get(uri=attribute_map['role'])
+                    for collection_index, role in enumerate(roles):
+                        r_affiliation = next(k for k,v in affiliations.items() if role in v)
+                        Value.objects.update_or_create(
+                            project=instance.project,
+                            attribute=attribute,
+                            set_prefix=instance.set_prefix,
+                            set_index=instance.set_index,
+                            collection_index=collection_index,
+                            external_id=r_affiliation,
+                            defaults={
+                                'text': role
+                            }
+                        )
+
+                    # delete surplus collection_indexes
+                    Value.objects.filter(
                         project=instance.project,
-                        attribute=attribute,
+                        snapshot=None,
                         set_prefix=instance.set_prefix,
                         set_index=instance.set_index,
-                        collection_index=collection_index,
-                        defaults={
-                            'text': affiliation
-                        }
-                    )
+                        attribute=attribute
+                    ).exclude(collection_index__in=range(len(roles))).delete()
 
-                # delete surplus collection_indexes
-                Value.objects.filter(
-                    project=instance.project,
-                    snapshot=None,
-                    set_prefix=instance.set_prefix,
-                    set_index=instance.set_index,
-                    attribute=attribute
-                ).exclude(collection_index__in=range(len(affiliations))).delete()
+                if 'affiliation' in attribute_map:
+                    attribute = Attribute.objects.get(uri=attribute_map['affiliation'])
+                    for collection_index, affiliation in enumerate(affiliations.keys()):
+                        role_string = ', '.join(affiliations[affiliation])
+                        Value.objects.update_or_create(
+                            project=instance.project,
+                            attribute=attribute,
+                            set_prefix=instance.set_prefix,
+                            set_index=instance.set_index,
+                            collection_index=collection_index,
+                            external_id=role_string,
+                            defaults={
+                                'text': affiliation
+                            }
+                        )
+
+                    # delete surplus collection_indexes
+                    Value.objects.filter(
+                        project=instance.project,
+                        snapshot=None,
+                        set_prefix=instance.set_prefix,
+                        set_index=instance.set_index,
+                        attribute=attribute
+                    ).exclude(collection_index__in=range(len(affiliations))).delete()
